Refactor AI prompt preview rendering in settings views
